requirements/base.txt
django==4.2.4
django-cors-headers==4.2.0
django-cachelot==2.6.1
django-redis==5.3.0
pillow==9.5.0
python-decouple==3.8